Change internal Emily code to use u16s instead of i32s

+++ emily/handler/src/api/models/deposit/requests.rs
@@ -14,7 +14,7 @@ pub struct GetDepositsForTransactionQuery {
     pub next_token: Option<String>,
     /// Maximum number of results to show.
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub page_size: Option<i32>,
+    pub page_size: Option<u16>,
 }
 
 /// Query structure for the GetDepositsQuery struct.
@@ -28,7 +28,7 @@ pub struct GetDepositsQuery {
     pub next_token: Option<String>,
     /// Maximum number of results to show.
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub page_size: Option<i32>,
+    pub page_size: Option<u16>,
 }
 
 /// Query structure common for all paginated queries.
@@ -40,7 +40,7 @@ pub struct BasicPaginationQuery {
     pub next_token: Option<String>,
     /// Maximum number of results to show.
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub page_size: Option<i32>,
+    pub page_size: Option<u16>,
 }
 
 /// Request structure for create deposit request.
