+++ emily/handler/src/api/models/withdrawal/requests.rs
@@ -17,7 +17,7 @@ pub struct GetWithdrawalsQuery {
     pub next_token: Option<String>,
     /// Maximum number of results to show.
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub page_size: Option<i32>,
+    pub page_size: Option<u16>,
 }
 
 /// Request structure for the create withdrawal request.
